The quoted-string production is one module, one walk
Ten items and nineteen tests, including the typed `QuotedStringDefect` and the
iterator the three entry points share. `validate_quoted_string` asks the walk
only whether it finished, `unescape_quoted_string` keeps what it produced, and a
test holds the two to that agreement — they were once separate walks.

The boundary against the list splitters is the interesting part, and the code
had already written it down in two places. `split_top_level`'s comment says the
productions "are transcribed at `validate_quoted_string`, which owns them", and
`quoting_is_balanced` says it "lives beside the splitters because it has to
agree with them about what a quote is". So the splitters stay where they are and
keep tracking quote parity themselves: `quoted-pair` is part of `quoted-string`
and nowhere else, and honouring an escape at top level once let a stray
backslash swallow every later member. Two productions that disagree, not one
duplicated.

Confirmation the cluster was cohesive: with it gone, `shown_in_finding` has no
remaining caller in `headers.rs` at all.

19 tests now run under `helpers::quoted_string`; crate total unchanged at 6032.
headers.rs: 2344 → 1803 non-test lines.
